rename: BridgeGame 패키지 변경
빌드 오류로 인하여 어떤 클래스가 문제인지 찾는중
